backend/requirements.txt
djangorestframework==3.11.2
django-extensions==2.1.4
requests==2.22.0
psycopg2-binary==2.8.2
graphene==2.1.3
graphene-django==2.4.0
dj-database-url==0.5.0
Django==2.2.14
django-redis==4.5.0
gunicorn==19.9.0
websocket-client==0.37.0
graphql-core==2.2.1
graphql-relay==0.4.5
dnspython==1.16.0
enum-compat==0.0.3
greenlet==0.4.16
monotonic==1.5
python-engineio
python-socketio
pytz==2020.1
six==1.15.0
ptyprocess
pytest-cov==2.8.1
pytest-django==3.5.1
pytest
pytest-xdist
pytest-sugar
pytest-vcr==1.0.2
django-stubs
djangorestframework-stubs
